+++ Pod/Classes/ViewControllers/ARShowComponentViewController.m
@@ -26,5 +26,16 @@
     return [super initWithEmission:emission moduleName:@"Show" initialProperties:@{ @"showID": showID }];
 }
 
+// Don't support rotation in iPhone
+
+- (UIInterfaceOrientationMask)supportedInterfaceOrientations
+{
+    return UI_USER_INTERFACE_IDIOM() != UIUserInterfaceIdiomPhone ? UIInterfaceOrientationMaskAll : UIInterfaceOrientationMaskAllButUpsideDown;
+}
+
+- (BOOL)shouldAutorotate
+{
+    return UI_USER_INTERFACE_IDIOM() != UIUserInterfaceIdiomPhone;
+}
 
 @end